Removes obsolete test cases
